+++ packages/firebase_auth/ios/Classes/FirebaseAuthPlugin.m
@@ -66,9 +66,10 @@ int nextHandle = 0;
           [auth removeAuthStateDidChangeListener:listener];
         }];
   } else if ([@"signInAnonymously" isEqualToString:call.method]) {
-    [[FIRAuth auth] signInAnonymouslyWithCompletion:^(FIRUser *user, NSError *error) {
-      [self sendResult:result forUser:user error:error];
-    }];
+    [[FIRAuth auth]
+        signInAnonymouslyWithCompletion:^(FIRAuthDataResult *dataResult, NSError *error) {
+          [self sendResult:result forUser:dataResult.user error:error];
+        }];
   } else if ([@"signInWithGoogle" isEqualToString:call.method]) {
     NSString *idToken = call.arguments[@"idToken"];
     NSString *accessToken = call.arguments[@"accessToken"];
@@ -85,13 +86,22 @@ int nextHandle = 0;
                               completion:^(FIRUser *user, NSError *error) {
                                 [self sendResult:result forUser:user error:error];
                               }];
+  } else if ([@"signInWithTwitter" isEqualToString:call.method]) {
+    NSString *authToken = call.arguments[@"authToken"];
+    NSString *authTokenSecret = call.arguments[@"authTokenSecret"];
+    FIRAuthCredential *credential =
+        [FIRTwitterAuthProvider credentialWithToken:authToken secret:authTokenSecret];
+    [[FIRAuth auth] signInWithCredential:credential
+                              completion:^(FIRUser *user, NSError *error) {
+                                [self sendResult:result forUser:user error:error];
+                              }];
   } else if ([@"createUserWithEmailAndPassword" isEqualToString:call.method]) {
     NSString *email = call.arguments[@"email"];
     NSString *password = call.arguments[@"password"];
     [[FIRAuth auth] createUserWithEmail:email
                                password:password
-                             completion:^(FIRUser *user, NSError *error) {
-                               [self sendResult:result forUser:user error:error];
+                             completion:^(FIRAuthDataResult *dataResult, NSError *error) {
+                               [self sendResult:result forUser:dataResult.user error:error];
                              }];
   } else if ([@"fetchProvidersForEmail" isEqualToString:call.method]) {
     NSString *email = call.arguments[@"email"];
@@ -103,7 +113,10 @@ int nextHandle = 0;
     [[FIRAuth auth].currentUser sendEmailVerificationWithCompletion:^(NSError *_Nullable error) {
       [self sendResult:result forProviders:nil error:error];
     }];
-
+  } else if ([@"reload" isEqualToString:call.method]) {
+    [[FIRAuth auth].currentUser reloadWithCompletion:^(NSError *_Nullable error) {
+      [self sendResult:result forProviders:nil error:error];
+    }];
   } else if ([@"sendPasswordResetEmail" isEqualToString:call.method]) {
     NSString *email = call.arguments[@"email"];
     [[FIRAuth auth] sendPasswordResetWithEmail:email
@@ -115,8 +128,8 @@ int nextHandle = 0;
     NSString *password = call.arguments[@"password"];
     [[FIRAuth auth] signInWithEmail:email
                            password:password
-                         completion:^(FIRUser *user, NSError *error) {
-                           [self sendResult:result forUser:user error:error];
+                         completion:^(FIRAuthDataResult *dataResult, NSError *error) {
+                           [self sendResult:result forUser:dataResult.user error:error];
                          }];
   } else if ([@"signOut" isEqualToString:call.method]) {
     NSError *signOutError;
@@ -172,8 +185,8 @@ int nextHandle = 0;
   } else if ([@"signInWithCustomToken" isEqualToString:call.method]) {
     NSString *token = call.arguments[@"token"];
     [[FIRAuth auth] signInWithCustomToken:token
-                               completion:^(FIRUser *user, NSError *error) {
-                                 [self sendResult:result forUser:user error:error];
+                               completion:^(FIRAuthDataResult *dataResult, NSError *error) {
+                                 [self sendResult:result forUser:dataResult.user error:error];
                                }];
 
   } else if ([@"startListeningAuthState" isEqualToString:call.method]) {