+++ packages/firebase_auth/lib/firebase_auth.dart
@@ -82,6 +82,11 @@ class FirebaseUser extends UserInfo {
     await FirebaseAuth.channel.invokeMethod('sendEmailVerification');
   }
 
+  /// Manually refreshes the data of the current user (for example, attached providers, display name, and so on).
+  Future<void> reload() async {
+    await FirebaseAuth.channel.invokeMethod('reload');
+  }
+
   @override
   String toString() {
     return '$runtimeType($_data)';
@@ -210,6 +215,24 @@ class FirebaseAuth {
     final Map<dynamic, dynamic> data =
         await channel.invokeMethod('signInWithFacebook', <String, String>{
       'accessToken': accessToken,
+    });
+    final FirebaseUser currentUser = new FirebaseUser._(data);
+    return currentUser;
+  }
+
+  /// Signs in with a Twitter account using the specified credentials.
+  ///
+  /// The returned future completes with the signed-in user or a [PlatformException], if sign in failed.
+  Future<FirebaseUser> signInWithTwitter({
+    @required String authToken,
+    @required String authTokenSecret,
+  }) async {
+    assert(authToken != null);
+    assert(authTokenSecret != null);
+    final Map<dynamic, dynamic> data =
+        await channel.invokeMethod('signInWithTwitter', <String, String>{
+      'authToken': authToken,
+      'authTokenSecret': authTokenSecret,
     });
     final FirebaseUser currentUser = new FirebaseUser._(data);
     return currentUser;