+++ packages/firebase_storage/test/firebase_storage_test.dart
@@ -19,6 +19,174 @@ void main() {
     final FirebaseStorage storage =
         new FirebaseStorage(app: app, storageBucket: storageBucket);
 
+    group('getMaxDownloadRetryTimeMillis', () {
+      final List<MethodCall> log = <MethodCall>[];
+
+      setUp(() {
+        FirebaseStorage.channel
+            .setMockMethodCallHandler((MethodCall methodCall) async {
+          log.add(methodCall);
+          return 1000;
+        });
+      });
+
+      test('invokes correct method', () async {
+        await storage.getMaxDownloadRetryTimeMillis();
+
+        expect(log, <Matcher>[
+          isMethodCall(
+            'FirebaseStorage#getMaxDownloadRetryTime',
+            arguments: <String, String>{
+              'app': 'testApp',
+              'bucket': 'gs://fake-storage-bucket-url.com',
+            },
+          ),
+        ]);
+      });
+
+      test('returns correct result', () async {
+        expect(await storage.getMaxDownloadRetryTimeMillis(), 1000);
+      });
+    });
+
+    group('getMaxUploadRetryTimeMillis', () {
+      final List<MethodCall> log = <MethodCall>[];
+
+      setUp(() {
+        FirebaseStorage.channel
+            .setMockMethodCallHandler((MethodCall methodCall) async {
+          log.add(methodCall);
+          return 2000;
+        });
+      });
+
+      test('invokes correct method', () async {
+        await storage.getMaxUploadRetryTimeMillis();
+
+        expect(log, <Matcher>[
+          isMethodCall(
+            'FirebaseStorage#getMaxUploadRetryTime',
+            arguments: <String, String>{
+              'app': 'testApp',
+              'bucket': 'gs://fake-storage-bucket-url.com',
+            },
+          ),
+        ]);
+      });
+
+      test('returns correct result', () async {
+        expect(await storage.getMaxUploadRetryTimeMillis(), 2000);
+      });
+    });
+
+    group('getMaxOperationRetryTimeMillis', () {
+      final List<MethodCall> log = <MethodCall>[];
+
+      setUp(() {
+        FirebaseStorage.channel
+            .setMockMethodCallHandler((MethodCall methodCall) async {
+          log.add(methodCall);
+          return 3000;
+        });
+      });
+
+      test('invokes correct method', () async {
+        await storage.getMaxOperationRetryTimeMillis();
+
+        expect(log, <Matcher>[
+          isMethodCall(
+            'FirebaseStorage#getMaxOperationRetryTime',
+            arguments: <String, String>{
+              'app': 'testApp',
+              'bucket': 'gs://fake-storage-bucket-url.com',
+            },
+          ),
+        ]);
+      });
+
+      test('returns correct result', () async {
+        expect(await storage.getMaxOperationRetryTimeMillis(), 3000);
+      });
+    });
+
+    group('setMaxDownloadRetryTimeMillis', () {
+      final List<MethodCall> log = <MethodCall>[];
+
+      setUp(() {
+        FirebaseStorage.channel
+            .setMockMethodCallHandler((MethodCall methodCall) async {
+          log.add(methodCall);
+        });
+      });
+
+      test('invokes correct method', () async {
+        await storage.setMaxDownloadRetryTimeMillis(1000);
+
+        expect(log, <Matcher>[
+          isMethodCall(
+            'FirebaseStorage#setMaxDownloadRetryTime',
+            arguments: <String, dynamic>{
+              'app': 'testApp',
+              'bucket': 'gs://fake-storage-bucket-url.com',
+              'time': 1000,
+            },
+          ),
+        ]);
+      });
+    });
+
+    group('setMaxUploadRetryTimeMillis', () {
+      final List<MethodCall> log = <MethodCall>[];
+
+      setUp(() {
+        FirebaseStorage.channel
+            .setMockMethodCallHandler((MethodCall methodCall) async {
+          log.add(methodCall);
+        });
+      });
+
+      test('invokes correct method', () async {
+        await storage.setMaxUploadRetryTimeMillis(2000);
+
+        expect(log, <Matcher>[
+          isMethodCall(
+            'FirebaseStorage#setMaxUploadRetryTime',
+            arguments: <String, dynamic>{
+              'app': 'testApp',
+              'bucket': 'gs://fake-storage-bucket-url.com',
+              'time': 2000,
+            },
+          ),
+        ]);
+      });
+    });
+
+    group('setMaxOperationRetryTimeMillis', () {
+      final List<MethodCall> log = <MethodCall>[];
+
+      setUp(() {
+        FirebaseStorage.channel
+            .setMockMethodCallHandler((MethodCall methodCall) async {
+          log.add(methodCall);
+        });
+      });
+
+      test('invokes correct method', () async {
+        await storage.setMaxOperationRetryTimeMillis(3000);
+
+        expect(log, <Matcher>[
+          isMethodCall(
+            'FirebaseStorage#setMaxOperationRetryTime',
+            arguments: <String, dynamic>{
+              'app': 'testApp',
+              'bucket': 'gs://fake-storage-bucket-url.com',
+              'time': 3000,
+            },
+          ),
+        ]);
+      });
+    });
+
     group('StorageReference', () {
       group('getData', () {
         final List<MethodCall> log = <MethodCall>[];
@@ -109,9 +277,10 @@ void main() {
                 };
                 break;
               case 'StorageReference#updateMetadata':
-                return <String, String>{
+                return <String, dynamic>{
                   'name': 'image.jpg',
-                  'contentLanguage': 'en'
+                  'contentLanguage': 'en',
+                  'customMetadata': <String, String>{'activity': 'test'},
                 };
                 break;
               default:
@@ -123,8 +292,10 @@ void main() {
         });
 
         test('invokes correct method', () async {
-          await ref
-              .updateMetadata(const StorageMetadata(contentLanguage: 'en'));
+          await ref.updateMetadata(new StorageMetadata(
+            contentLanguage: 'en',
+            customMetadata: <String, String>{'activity': 'test'},
+          ));
 
           expect(log, <Matcher>[
             isMethodCall(
@@ -133,12 +304,13 @@ void main() {
                 'app': 'testApp',
                 'bucket': 'gs://fake-storage-bucket-url.com',
                 'path': 'avatars/large/image.jpg',
-                'metadata': <String, String>{
+                'metadata': <String, dynamic>{
                   'cacheControl': null,
                   'contentDisposition': null,
                   'contentLanguage': 'en',
                   'contentType': null,
-                  'contentEncoding': null
+                  'contentEncoding': null,
+                  'customMetadata': <String, String>{'activity': 'test'},
                 },
               },
             ),
@@ -149,7 +321,7 @@ void main() {
           expect((await ref.getMetadata()).contentLanguage, null);
           expect(
               (await ref.updateMetadata(
-                      const StorageMetadata(contentLanguage: 'en')))
+                      new StorageMetadata(contentLanguage: 'en')))
                   .contentLanguage,
               'en');
         });