+++ packages/sensors/lib/sensors.dart
@@ -1,12 +1,14 @@
 import 'dart:async';
-
 import 'package:flutter/services.dart';
 
 const EventChannel _accelerometerEventChannel =
-    const EventChannel('plugins.flutter.io/accelerometer');
+    const EventChannel('plugins.flutter.io/sensors/accelerometer');
+
+const EventChannel _userAccelerometerEventChannel =
+    const EventChannel('plugins.flutter.io/sensors/user_accel');
 
 const EventChannel _gyroscopeEventChannel =
-    const EventChannel('plugins.flutter.io/gyroscope');
+    const EventChannel('plugins.flutter.io/sensors/gyroscope');
 
 class AccelerometerEvent {
   /// Acceleration force along the x axis (including gravity) measured in m/s^2.
@@ -40,8 +42,28 @@ class GyroscopeEvent {
   String toString() => '[GyroscopeEvent (x: $x, y: $y, z: $z)]';
 }
 
+class UserAccelerometerEvent {
+  /// Acceleration force along the x axis (excluding gravity) measured in m/s^2.
+  final double x;
+
+  /// Acceleration force along the y axis (excluding gravity) measured in m/s^2.
+  final double y;
+
+  /// Acceleration force along the z axis (excluding gravity) measured in m/s^2.
+  final double z;
+
+  UserAccelerometerEvent(this.x, this.y, this.z);
+
+  @override
+  String toString() => '[UserAccelerometerEvent (x: $x, y: $y, z: $z)]';
+}
+
 AccelerometerEvent _listToAccelerometerEvent(List<double> list) {
   return new AccelerometerEvent(list[0], list[1], list[2]);
+}
+
+UserAccelerometerEvent _listToUserAccelerometerEvent(List<double> list) {
+  return new UserAccelerometerEvent(list[0], list[1], list[2]);
 }
 
 GyroscopeEvent _listToGyroscopeEvent(List<double> list) {
@@ -50,6 +72,7 @@ GyroscopeEvent _listToGyroscopeEvent(List<double> list) {
 
 Stream<AccelerometerEvent> _accelerometerEvents;
 Stream<GyroscopeEvent> _gyroscopeEvents;
+Stream<UserAccelerometerEvent> _userAccelerometerEvents;
 
 /// A broadcast stream of events from the device accelerometer.
 Stream<AccelerometerEvent> get accelerometerEvents {
@@ -69,4 +92,14 @@ Stream<GyroscopeEvent> get gyroscopeEvents {
         .map((dynamic event) => _listToGyroscopeEvent(event));
   }
   return _gyroscopeEvents;
+}
+
+/// Events from the device accelerometer with gravity removed.
+Stream<UserAccelerometerEvent> get userAccelerometerEvents {
+  if (_userAccelerometerEvents == null) {
+    _userAccelerometerEvents = _userAccelerometerEventChannel
+        .receiveBroadcastStream()
+        .map((dynamic event) => _listToUserAccelerometerEvent(event));
+  }
+  return _userAccelerometerEvents;
 }