+++ packages/cloud_firestore/CHANGELOG.md
@@ -1,3 +1,26 @@
+## 0.7.2
+
+* Fixes crash on Android if a FirebaseFirestoreException happened.
+
+## 0.7.1
+
+* Updated iOS implementation to reflect Firebase API changes.
+* Fixed bug in Transaction.get that would fail on no data.
+* Fixed error in README.md code sample.
+
+## 0.7.0+2
+
+* Update transactions example in README to add `await`.
+
+## 0.7.0+1
+
+* Add transactions example to README.
+
+## 0.7.0
+
+* **Breaking change**. `snapshots` is now a method instead of a getter.
+* **Breaking change**. `setData` uses named arguments instead of `SetOptions`.
+
 ## 0.6.3
 
 * Updated Google Play Services dependencies to version 15.0.0.