+++ packages/firebase_remote_config/CHANGELOG.md
@@ -1,3 +1,16 @@
+## [0.0.3]
+
+* Added missing await in setDefaults.
+* Fixed example code in README.
+
+## [0.0.2]
+
+* Update iOS plugin so that it returns fetch status
+  as a String instead of an int.
+* Bump Android library version to 15.+. The Android plugins for
+  FlutterFire need to all be on the same version. Updating
+  Remote Config to match other FlutterFire plugins.
+
 ## [0.0.1]
 
 * Implement Firebase Remote Config.