+++ packages/firebase_storage/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.3.6
+
+* Added support for custom metadata.
+
+## 0.3.5
+
+* Updated iOS implementation to reflect Firebase API changes.
+
+## 0.3.4
+
+* Added timeout properties to FirebaseStorage.
+
 ## 0.3.3
 
 * Added support for initialization with a custom Firebase app.